Gather all element metadata before upgrading scripts.
Element metadata includes stuff like attributes, host attributes,
new computed functions. We need this metadata to build out the Polymer()
declaration for the element.

Before we were upgrading scripts more or less as we saw them, and not
keeping track of all element metadata in one place. Now, we first gather
all of the metadata, so that we'll be able to upgrade any Polymer() call
for any element in any <script>. The most common real world case that this
improves is when a script contains more than one call to Polymer().